Version 0.15.1 (Beta 15.1)
Release Notes
==========
* Added a check system to the server configuration builder to see if
there are key fields missing from the plugin prefs.  This is in direct
response to two users reporting that their "apiport" was missing
([Issue
#12](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/12))
* Added ability for contact sensors, which already natively invert the
on state of a device, to utilize the "invert" checkbox so that it can
**not** invert it instead, this in direct response to [Issue
#32](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/32)
* Added new utility to the Advanced Plugin Options (under the plugin
menu) to [Simulate a HomeKit
device](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/Plugin-M
enu-Utilities#advanced-plugin-actions) to better diagnose what may be
going on, this allows you to see what HomeKit will see when your device
is sent to it
* Removed all complications implementation from the plugin as it's
going to be totally different when implemented than it was in the
experimental implementation, yet the experimental settings were still
in the plugin and causing confusion
* Changed
[Faucet](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/HomeKit
-Model-Reference#faucet) to be 3rd Party Only since in Home it
constantly says "Updating".  You can use
[Valve](https://github.com/Colorado4Wheeler/HomeKit-Bridge/wiki/HomeKit-
Model-Reference#valve) instead if you want that icon and more
functionality.  This in direct response to [Issue
#38](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/38)
* Fixed bug where the plugin was loading an unneeded library and
**might** cause an error if it ran across a plugin that didn't have an
info.plist file
* Fixed bug where choosing a device that could not be auto detected
(for instance a "custom" plugin device) then it would generate an error
about being able to "iteritems".  Now non detectable devices will
default to a switch and a warning will show in the log that it could
not be figured out.
* [Issue
#32](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/32)
resolved
* [Issue
#38](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/38)
resolved