Version 0.16.3 (Beta 16.3)
Release Notes
==========
* **NOTE: Due to a fundamental change in temp control you will need to
edit all of your sensors and thermostats and determine if the new
"Temperature value is Fahrenheit" needs to be checks so the plugin
knows to convert it to Celsius.  It's possible that when you add a
temperature sensor or thermostat to a server that you won't see the
checkbox, I believe this is only because it's an existing server, if
you add it and there is no checkbox it'll be there when you edit that
same item.**
* Ongoing update to 16.1 and 16.2 bugs to try to finally close ticket 37
* [Issue
#37](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/37)
hopefully resolved now!