Version 0.21.0 (Beta 21.0)
* **NOTE** This release begins the process of rewriting the entire
HomeKit integration, small chunks at a time, so it's possible that this
can break things.  I'm trying to thoroughly test changes to make sure
the end result is the same, but you have been warned.
* Added failsafe when saving the Homebridge configuration that if there
are no Indigo devices being passed that the Homebridge-indigo2 platform
will not be loaded, this in case you are using a "camera server" that
has nothing else on it doesn't cause issues with Homebridge-Indigo2
trying to load zero accessories
* Changed model:submodel device resolution to detect if either model or
submodel is an empty attribute and not pass the ":" separator if that
is the case
* Completely rewrote the SecuritySpy URL construction routine to better
support users who don't store the xpassword state ([Issue
#81](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/81))
* Moved model/firmware resolution out of the main plugin and into the
new factory package
* Code housekeeping and cleanup