Version 0.22.1 (Beta 22.1)
* **NOTE** As stated earlier, the entire HomeKit engine is being
optimized and rewritten for various reasons.  There is a new plugin
configuration option that allows you to revert to the previous method
if the new method is causing problem, but this is on a
release-by-release basis, meaning that only changes in **this** release
will roll back to using the old methods, the next release will not be
able to roll back anything from this release.  The functions impacted
will be noted in the release notes as 'Library Change'.  Please report
any issues that are caused by the new library that are resolved by
returning to the old methods.  Each revision will re-enable this option
automatically if it gets turned off.
* **NOTE** All minor revisions can still be rolled back for Library
Changes until the next full beta version is released
* Added optimized Homebridge callback POST method to replace the
current GET method, in preparation for an updated Homebridge-Indigo2
script that will parse the data instead of calling back to the API
([Issue
#93]((https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/93))
* Fixed bug in configuration builder where if an error was tripped the
exception would not populate ([Issue
#94]((https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/94) and
[Issue
#92]((https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/92))
* Fixed bug in new library that would allow Homebridge to see a
duplicate camera device and would send that device to HomeKit along
with the normal camera ([Issue
#83](https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/83))
* [Issue
#94]((https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/94)
resolved
* [Issue
#92]((https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/92)
resolved
* [Issue
#93]((https://github.com/Colorado4Wheeler/HomeKit-Bridge/issues/93)
implemented